+++ src/NativeStripeSdk.tsx
@@ -11,6 +11,7 @@ import type {
   RetrieveSetupIntentResult,
   ConfirmPaymentResult,
   HandleNextActionResult,
+  HandleNextActionForSetupResult,
   ConfirmSetupIntentResult,
   CreateTokenForCVCUpdateResult,
   InitPaymentSheetResult,
@@ -36,6 +37,10 @@ type NativeStripeSdkType = {
     paymentIntentClientSecret: string,
     returnURL?: string | null
   ): Promise<HandleNextActionResult>;
+  handleNextActionForSetup(
+    setupIntentClientSecret: string,
+    returnURL?: string | null
+  ): Promise<HandleNextActionForSetupResult>;
   confirmPayment(
     paymentIntentClientSecret: string,
     params?: PaymentIntent.ConfirmParams,