+++ src/hooks/useStripe.tsx
@@ -7,6 +7,7 @@ import type {
   RetrieveSetupIntentResult,
   ConfirmPaymentResult,
   HandleNextActionResult,
+  HandleNextActionForSetupResult,
   ConfirmSetupIntentResult,
   CreateTokenForCVCUpdateResult,
   StripeError,
@@ -36,6 +37,7 @@ import {
   confirmSetupIntent,
   createTokenForCVCUpdate,
   handleNextAction,
+  handleNextActionForSetup,
   handleURLCallback,
   initPaymentSheet,
   presentPaymentSheet,
@@ -111,6 +113,16 @@ export function useStripe() {
       returnURL?: string
     ): Promise<HandleNextActionResult> => {
       return handleNextAction(paymentIntentClientSecret, returnURL);
+    },
+    []
+  );
+
+  const _handleNextActionForSetup = useCallback(
+    async (
+      setupIntentClientSecret: string,
+      returnURL?: string
+    ): Promise<HandleNextActionForSetupResult> => {
+      return handleNextActionForSetup(setupIntentClientSecret, returnURL);
     },
     []
   );
@@ -308,6 +320,7 @@ export function useStripe() {
     confirmPayment: _confirmPayment,
     createPaymentMethod: _createPaymentMethod,
     handleNextAction: _handleNextAction,
+    handleNextActionForSetup: _handleNextActionForSetup,
     confirmSetupIntent: _confirmSetupIntent,
     createTokenForCVCUpdate: _createTokenForCVCUpdate,
     handleURLCallback: _handleURLCallback,